server/requirements.txt
aiofiles==0.4.0
alembic==1.7.5
asgiref==3.4.1
asyncpg==0.25.0
attrs==21.2.0
azure-core==1.20.1
azure-identity==1.7.1
azure-storage-blob==12.7.1
black==21.11b1
blacksheep==1.2.1
blacksheep-sqlalchemy==0.0.2
cachetools==4.2.4
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
coverage==6.2
cryptography==36.0.0
essentials==1.1.4
essentials-configuration==0.0.2
essentials-openapi==0.1.6
flake8==4.0.1
gallerist==0.0.5
gallerist-azurestorage==0.0.4
google-api-core==2.2.2
google-auth==2.3.3
googleapis-common-protos==1.53.0
greenlet==1.1.2
guardpost==0.0.9
h11==0.12.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
isodate==0.6.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
mypy-extensions==0.4.3
oauthlib==3.1.1
opencensus==0.8.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.0
opencensus-ext-logging==0.1.1
opencensus-ext-sqlalchemy==0.1.2
packaging==21.3
pathspec==0.9.0
Pillow==8.1.2
platformdirs==2.4.0
pluggy==1.0.0
portalocker==1.7.1
protobuf==3.19.1
psutil==5.8.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-cov==3.0.0
python-dateutil==2.8.2
python-slugify==5.0.2
PyYAML==5.4.1
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rodi==1.1.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.15
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
zipp==3.6.0
